+++ about/src/main/kotlin/ca/allanwang/kau/about/AboutActivityBase.kt
@@ -13,7 +13,8 @@ import android.view.ViewGroup
 import ca.allanwang.kau.adapters.FastItemThemedAdapter
 import ca.allanwang.kau.adapters.ThemableIItemColors
 import ca.allanwang.kau.adapters.ThemableIItemColorsDelegate
-import ca.allanwang.kau.animators.FadeScaleAnimator
+import ca.allanwang.kau.animators.FadeScaleAnimatorAdd
+import ca.allanwang.kau.animators.KauAnimator
 import ca.allanwang.kau.iitems.HeaderIItem
 import ca.allanwang.kau.ui.widgets.ElasticDragDismissFrameLayout
 import ca.allanwang.kau.ui.widgets.InkPageIndicator
@@ -94,12 +95,9 @@ abstract class AboutActivityBase(val rClass: Class<*>?, val configBuilder: Confi
         indicator.setViewPager(pager)
         draggableFrame.addListener(object : ElasticDragDismissFrameLayout.SystemChromeFader(this) {
             override fun onDragDismissed() {
-                // if we drag dismiss downward then the default reversal of the enter
-                // transition would slide content upward which looks weird. So reverse it.
-                if (draggableFrame.translationY > 0) {
-                    window.returnTransition = TransitionInflater.from(this@AboutActivityBase)
-                            .inflateTransition(configs.transitionExitReversed)
-                }
+                window.returnTransition = TransitionInflater.from(this@AboutActivityBase)
+                        .inflateTransition(if (draggableFrame.translationY > 0) configs.transitionExitBottom else configs.transitionExitTop)
+
                 libRecycler?.stopScroll()
                 finishAfterTransition()
             }
@@ -114,10 +112,9 @@ abstract class AboutActivityBase(val rClass: Class<*>?, val configBuilder: Confi
         var cutoutForeground: Int? = null
         var libPageTitleRes: Int = -1
         var libPageTitle: String? = string(R.string.kau_about_libraries_intro) //This is in the string by default since it's lower priority
-        /**
-         * Transition to be called if the view is dragged down
-         */
-        var transitionExitReversed: Int = R.transition.kau_about_return_downward
+
+        var transitionExitTop: Int = R.transition.kau_exit_slide_top
+        var transitionExitBottom: Int = R.transition.kau_exit_slide_bottom
     }
 
     /**
@@ -174,8 +171,9 @@ abstract class AboutActivityBase(val rClass: Class<*>?, val configBuilder: Confi
         val v = layoutInflater.inflate(R.layout.kau_recycler_detached_background, parent, false)
         val recycler = v.findViewById<RecyclerView>(R.id.kau_recycler_detached)
         libRecycler = recycler
+        recycler.withMarginDecoration(16, KAU_BOTTOM)
         recycler.adapter = libAdapter
-        recycler.itemAnimator = FadeScaleAnimator(itemDelayFactor = 0.2f).apply { addDuration = 300; interpolator = AnimHolder.decelerateInterpolator(this@AboutActivityBase) }
+        recycler.itemAnimator = KauAnimator(addAnimator = FadeScaleAnimatorAdd(scaleFactor = 0.7f, itemDelayFactor = 0.2f)).apply { addDuration = 300; interpolator = AnimHolder.decelerateInterpolator(this@AboutActivityBase) }
         val background = v.findViewById<View>(R.id.kau_recycler_detached_background)
         if (configs.backgroundColor != null) background.setBackgroundColor(configs.backgroundColor!!.colorToForeground())
         doAsync {