+++ imagepicker/src/main/kotlin/ca/allanwang/kau/imagepicker/ImagePickerActivityBase.kt
@@ -1,5 +1,7 @@
 package ca.allanwang.kau.imagepicker
 
+import android.Manifest
+import android.content.Context
 import android.database.Cursor
 import android.os.Bundle
 import android.provider.MediaStore
@@ -10,9 +12,10 @@ import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.GridLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.support.v7.widget.Toolbar
-import ca.allanwang.kau.logging.KL
+import ca.allanwang.kau.permissions.kauRequestPermissions
 import ca.allanwang.kau.ui.widgets.ElasticDragDismissFrameLayout
 import ca.allanwang.kau.utils.bindView
+import ca.allanwang.kau.utils.dimenPixelSize
 import com.mikepenz.fastadapter.commons.adapters.FastItemAdapter
 
 
@@ -27,14 +30,37 @@ abstract class ImagePickerActivityBase : AppCompatActivity(), LoaderManager.Load
     val recycler: RecyclerView by bindView(R.id.kau_recycler)
     val imageAdapter = FastItemAdapter<ImageItem>()
 
+    companion object {
+        /**
+         * Given the dimensions of our device and a minimum image size,
+         * Computer the optimal column count for our grid layout
+         *
+         * @return column count
+         */
+        fun computeColumnCount(context: Context): Int {
+            val minImageSizePx = context.dimenPixelSize(R.dimen.kau_image_minimum_size)
+            val screenWidthPx = context.resources.displayMetrics.widthPixels
+            return screenWidthPx / minImageSizePx
+        }
+
+        var accentColor: Int = 0xff666666.toInt()
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.kau_activity_image_picker)
-        with(recycler) {
-            layoutManager = GridLayoutManager(this@ImagePickerActivityBase, 3)
-            adapter = this@ImagePickerActivityBase.imageAdapter
+        recycler.layoutManager = GridLayoutManager(this, computeColumnCount(this))
+        recycler.adapter = imageAdapter
+
+        with(imageAdapter) {
+            withPositionBasedStateManagement(false)
+            withMultiSelect(true)
+            withSelectable(true)
+            withOnClickListener { v, _, _, _ ->
+                (v as BlurredImageView).toggleBlur()
+                true
+            }
         }
-        imageAdapter.add(arrayOf("a", "b", "c").map { ImageItem(it) })
         draggableFrame.addListener(object : ElasticDragDismissFrameLayout.SystemChromeFader(this) {
             override fun onDragDismissed() {
                 if (draggableFrame.translationY < 0) {
@@ -44,33 +70,45 @@ abstract class ImagePickerActivityBase : AppCompatActivity(), LoaderManager.Load
                 finishAfterTransition()
             }
         })
+        kauRequestPermissions(Manifest.permission.READ_EXTERNAL_STORAGE) {
+            granted, _ ->
+            if (granted) {
+                supportLoaderManager.initLoader(42, null, this)
+            }
+        }
     }
 
     override fun onCreateLoader(id: Int, args: Bundle?): Loader<Cursor> {
         val columns = arrayOf(
                 MediaStore.Images.Media._ID,
+                MediaStore.Images.Media.TITLE,
                 MediaStore.Images.Media.DATA,
+                MediaStore.Images.Media.SIZE,
                 MediaStore.Images.Media.DISPLAY_NAME,
-                MediaStore.Images.Media.DATE_MODIFIED)
-
-        return CursorLoader(this, MediaStore.Images.Media.EXTERNAL_CONTENT_URI, columns, null, null, null)
+                MediaStore.Images.Media.DATE_MODIFIED
+        )
+        return CursorLoader(this,
+                MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
+                columns,
+                null,
+                null,
+                MediaStore.Images.Media.DATE_MODIFIED + " DESC")
     }
 
-    override fun onLoadFinished(loader: Loader<Cursor>, data: Cursor) {
-        val dataIndex = data.getColumnIndex(MediaStore.Images.Media.DATA)
-        val alstPhotos = mutableListOf<String>()
-
-        data.moveToLast()
-        while (!data.isBeforeFirst) {
-            val photoPath = data.getString(dataIndex)
-            KL.d(photoPath)
-            alstPhotos.add(photoPath)
-            data.moveToPrevious()
+    override fun onLoadFinished(loader: Loader<Cursor>, data: Cursor?) {
+        reset()
+        if (data == null) return
+        if (data.moveToFirst()) {
+            do {
+                val img = ImageModel(data)
+                imageAdapter.add(ImageItem(img))
+            } while (data.moveToNext())
         }
-        imageAdapter.add(alstPhotos.map { ImageItem(it) })
     }
 
-    override fun onLoaderReset(loader: Loader<Cursor>) {
-        imageAdapter.clear()
+    private fun reset() {
+        imageAdapter.clear();
     }
+
+    override fun onLoaderReset(loader: Loader<Cursor>) = reset()
 }