+++ sample/src/main/kotlin/ca/allanwang/kau/sample/MainActivity.kt
@@ -154,12 +154,16 @@ class MainActivity : KPrefActivity() {
             descRes = R.string.sub_item_desc
         }
 
-        plainText(R.string.kau_lorem_ipsum) {
-            onClick = {
-                _, _, _ ->
-                startActivity(AboutActivity::class.java, transition = true)
-                false
-            }
+        plainText(R.string.gallery_showcase) {
+            onClick = { _, _, _ -> startActivity(ImageActivity::class.java, transition = true); false }
+        }
+
+        plainText(R.string.adapter_showcase) {
+            onClick = { _, _, _ -> startActivity(AdapterActivity::class.java, transition = true); false }
+        }
+
+        plainText(R.string.kau_about_app) {
+            onClick = { _, _, _ -> startActivity(AboutActivity::class.java, transition = true); false }
         }
 
     }
@@ -215,7 +219,7 @@ class MainActivity : KPrefActivity() {
             noResultsFound = R.string.kau_no_results_found
             shouldClearOnClose = false
             onItemClick = {
-                position, key, content, searchView ->
+                _, _, content, searchView ->
                 toast(content)
                 searchView.revealClose()
             }